+++ core/src/test/java/org/elasticsearch/discovery/AbstractDisruptionTestCase.java
@@ -51,7 +51,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -97,7 +96,7 @@ public abstract class AbstractDisruptionTestCase extends ESIntegTestCase {
     @Override
     public void setDisruptionScheme(ServiceDisruptionScheme scheme) {
         if (scheme instanceof NetworkDisruption &&
-                ((NetworkDisruption) scheme).getNetworkLinkDisruptionType() instanceof NetworkUnresponsive) {
+                ((NetworkDisruption) scheme).getNetworkLinkDisruptionType() instanceof NetworkDisruption.NetworkUnresponsive) {
             // the network unresponsive disruption may leave operations in flight
             // this is because this disruption scheme swallows requests by design
             // as such, these operations will never be marked as finished
@@ -186,80 +185,6 @@ public abstract class AbstractDisruptionTestCase extends ESIntegTestCase {
         }
     }
 
-    /**
-     * Simulates an unresponsive target node by dropping requests sent from source to target node.
-     */
-    public static class NetworkUnresponsive extends NetworkLinkDisruptionType {
-
-        @Override
-        public void applyDisruption(MockTransportService sourceTransportService, MockTransportService targetTransportService) {
-            sourceTransportService.addUnresponsiveRule(targetTransportService);
-        }
-
-        @Override
-        public String toString() {
-            return "network unresponsive";
-        }
-    }
-
-    /**
-     * Simulates slow or congested network. Delivery of requests that are sent from source to target node are delayed by a configurable
-     * time amount.
-     */
-    public static class NetworkDelay extends NetworkLinkDisruptionType {
-
-        public static TimeValue DEFAULT_DELAY_MIN = TimeValue.timeValueSeconds(10);
-        public static TimeValue DEFAULT_DELAY_MAX = TimeValue.timeValueSeconds(90);
-
-        private final TimeValue delay;
-
-        /**
-         * Delays requests by a fixed time value.
-         *
-         * @param delay time to delay requests
-         */
-        public NetworkDelay(TimeValue delay) {
-            this.delay = delay;
-        }
-
-        /**
-         * Delays requests by a random but fixed time value between {@link #DEFAULT_DELAY_MIN} and {@link #DEFAULT_DELAY_MAX}.
-         *
-         * @param random instance to use for randomization of delay
-         */
-        public static NetworkDelay random(Random random) {
-            return random(random, DEFAULT_DELAY_MIN, DEFAULT_DELAY_MAX);
-        }
-
-        /**
-         * Delays requests by a random but fixed time value between delayMin and delayMax.
-         *
-         * @param random   instance to use for randomization of delay
-         * @param delayMin minimum delay
-         * @param delayMax maximum delay
-         */
-        public static NetworkDelay random(Random random, TimeValue delayMin, TimeValue delayMax) {
-            return new NetworkDelay(TimeValue.timeValueMillis(delayMin.millis() == delayMax.millis() ?
-                    delayMin.millis() :
-                    delayMin.millis() + random.nextInt((int) (delayMax.millis() - delayMin.millis()))));
-        }
-
-        @Override
-        public void applyDisruption(MockTransportService sourceTransportService, MockTransportService targetTransportService) {
-            sourceTransportService.addUnresponsiveRule(targetTransportService, delay);
-        }
-
-        @Override
-        public TimeValue expectedTimeToHeal() {
-            return delay;
-        }
-
-        @Override
-        public String toString() {
-            return "network delays for [" + delay + "]";
-        }
-    }
-
     ClusterState getNodeClusterState(String node) {
         return client(node).admin().cluster().prepareState().setLocal(true).get().getState();
     }
@@ -329,13 +254,13 @@ public abstract class AbstractDisruptionTestCase extends ESIntegTestCase {
         final NetworkLinkDisruptionType disruptionType;
         switch (randomInt(2)) {
             case 0:
-                disruptionType = new NetworkUnresponsive();
+                disruptionType = new NetworkDisruption.NetworkUnresponsive();
                 break;
             case 1:
                 disruptionType = new NetworkDisconnect();
                 break;
             case 2:
-                disruptionType = NetworkDelay.random(random());
+                disruptionType = NetworkDisruption.NetworkDelay.random(random());
                 break;
             default:
                 throw new IllegalArgumentException();
@@ -353,7 +278,7 @@ public abstract class AbstractDisruptionTestCase extends ESIntegTestCase {
     NetworkDisruption addRandomDisruptionType(TwoPartitions partitions) {
         final NetworkLinkDisruptionType disruptionType;
         if (randomBoolean()) {
-            disruptionType = new NetworkUnresponsive();
+            disruptionType = new NetworkDisruption.NetworkUnresponsive();
         } else {
             disruptionType = new NetworkDisconnect();
         }
